Revert "LPS-115615 Rename"
This reverts commit 3687a4b45a41837a2264ad33de62e0446bebf828.
